requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
dj-database-url==3.0.1
Django==4.2.4
django-heroku==0.3.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
idna==3.10
mysqlclient==2.2.7
numpy==2.2.6
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psycopg2==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.31.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
